Modscript PacketOrder data change (beta). Updated Minimap Widget IDs. Added doClick (doAction), walkTileViewport (doAction), and walkTileMinimap (packet emulation; beta) to api. Updated packet method predicate desc. Beta packet emulation implemented (minimapWalk only atm). Improved DisableRendering to include tiles/entire region, and all renderable models (doAction scripts only; reduced cpu+memory).

+++ src/org/osrs/api/wrappers/PacketContext.java
@@ -15,5 +15,5 @@ public interface PacketContext{
 	public IncomingPacketMeta lastIncomingPacket();
 	public int idleReadPulses();
 	public IncomingPacketMeta thirdIncomingPacket();
-	public void sendPacket(OutgoingPacket packet, int a);
+	public void sendPacket(OutgoingPacket packet, byte a);
 }
